+++ src/flake8_logging/__init__.py
@@ -7,6 +7,8 @@ from functools import cache
 from importlib.metadata import version
 from typing import Any, cast
 
+from flake8_logging.root_logger import RootLoggerVisitor
+
 
 class Plugin:
     name = "flake8-logging"
@@ -22,6 +24,11 @@ class Plugin:
         type_ = type(self)
         for line, col, msg in visitor.errors:
             yield line, col, msg, type_
+
+        root_visitor = RootLoggerVisitor(logger_methods)
+        root_visitor.visit(self._tree)
+        for line, col in root_visitor.errors:
+            yield line, col, LOG016, type_
 
 
 logger_methods = frozenset(
@@ -126,7 +133,7 @@ LOG012 = "LOG012 formatting error: {n} {style} placeholder{ns} but {m} argument{
 LOG013 = "LOG013 formatting error: {mistake} key{ns}: {keys}"
 LOG014 = "LOG014 avoid exc_info=True outside of exception handlers"
 LOG015 = "LOG015 avoid logging calls on the root logger"
-LOG016 = "LOG016 avoid implicitly getting the root logger"
+LOG016 = "LOG016 avoid logging through an implicitly obtained root logger"
 
 
 class Visitor(ast.NodeVisitor):
@@ -211,9 +218,6 @@ class Visitor(ast.NodeVisitor):
             and node.func.id == "getLogger"
             and self._from_imports.get("getLogger") == "logging"
         ):
-            if not node.args and not node.keywords:
-                self.errors.append((node.lineno, node.col_offset, LOG016))
-
             if (
                 len(self._stack) >= 2
                 and isinstance(assign := self._stack[-2], ast.Assign)